Add raytracer to gui.
Moving in and out is messed up, but some bugs were fixed relating to
this.

1. When a ray's origin is within the box, and we are finding an initial
point, cast ray backwords. TODO: try this at other levels, we are
skipping depths > 50 (out of 100).

2. Don't count a valid hit behind the origin.

3. hit_BoundingBox had to be sure to return something in /entrance/,
which was not happening when the origin was within the bounding box.

+++ util.h
@@ -6,8 +6,8 @@
 
 #define UFor( i, bel )  for (i = 0; i < bel; ++i)
 
-    /* typedef double real; */
-typedef float real;
+typedef double real;
+    /* typedef float real; */
 typedef unsigned uint;
 typedef char tristate;
 typedef unsigned char byte;
